Refactor: refactored old workerpools
