+++ core/objectstorage/test/objectstorage_test.go
@@ -14,6 +14,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
+	"github.com/iotaledger/hive.go/runtime/workerpool"
+
 	"github.com/iotaledger/hive.go/core/kvstore"
 	"github.com/iotaledger/hive.go/core/kvstore/badger"
 	"github.com/iotaledger/hive.go/core/kvstore/mapdb"
@@ -22,7 +24,6 @@ import (
 	"github.com/iotaledger/hive.go/core/testutil"
 	"github.com/iotaledger/hive.go/core/types"
 	"github.com/iotaledger/hive.go/core/typeutils"
-	"github.com/iotaledger/hive.go/core/workerpool"
 )
 
 const (
@@ -81,7 +82,7 @@ func TestConcurrentCreateDelete(t *testing.T) {
 	metadataStorage := objectstorage.New(badgerDBMetadataStorage, testObjectFactory)
 
 	// create workerpool
-	wp := workerpool.NewBlockingQueuedWorkerPool(workerpool.WorkerCount(1024), workerpool.QueueSize(objectCount), workerpool.FlushTasksAtShutdown(true))
+	wp := workerpool.New(t.Name(), 1024)
 	wp.Start()
 
 	// result counters
@@ -126,7 +127,7 @@ func TestConcurrentCreateDelete(t *testing.T) {
 	}
 
 	// wait for workers to finish
-	wp.StopAndWait()
+	wp.Shutdown(false)
 
 	// count messages still in the store
 	messagesInStore := 0
